Revert "revert d4aecec6"
This reverts commit ffdded842fff09aee95482c7e4eb078c25d0c67a.

+++ libraries/chain/wasm_price.cpp
@@ -517,7 +517,7 @@ namespace contento { namespace chain { namespace wasm_price {
     // contract_balance_api
     WASM_PRICE_FN_CONSTS(
                          (get_contract_balance_internal,      (void*, int) )
-                         (transfer,                  (void*,int64_t, int))
+                         (transfer,                  (void*,int, int))
                          (get_value_internal,                 (int64_t))
                          (accept_pay,            (void*))
                          );
